% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{mplus_to_dsem}
\alias{mplus_to_dsem}
\title{Translate supported Mplus-like syntax}
\usage{
mplus_to_dsem(
  model,
  id = NULL,
  time = NULL,
  family = c("gaussian", "binary", "ordinal")
)
}
\arguments{
\item{model}{Character Mplus-like model input.}

\item{id, time, family}{Passed to \code{\link[=dsem_model]{dsem_model()}}.}
}
\value{
A compiled \code{DSEMmodel} using the same canonical representation as
lavaan-style input.
}
\description{
Translate supported Mplus-like syntax
}
