% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{dsem_inspect}
\alias{dsem_inspect}
\title{Inspect a DSEMr object}
\usage{
dsem_inspect(
  object,
  what = c("model", "terms", "estimates", "draws", "diagnostics", "compute", "versions",
    "fit")
)
}
\arguments{
\item{object}{A \code{DSEMmodel} or \code{DSEMfit}.}

\item{what}{Component to extract.}
}
\value{
The requested component.
}
\description{
Inspect a DSEMr object
}
